CRUD ClassRoom

+++ app/Http/Controllers/Grades/GradesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Grades;
 
 use App\Http\Controllers\Controller;
 use App\Models\Grade;
+use App\Models\Classroom;
 use App\Http\Requests\StoreGrades;
 use Illuminate\Http\Request;
 
@@ -12,9 +13,7 @@ class GradesController extends Controller
 
     public function index()
     {
-
         $Grades = Grade::all();
-
         return view('pages.Grades.Grades' , compact('Grades')) ;
     }
 
@@ -28,6 +27,7 @@ class GradesController extends Controller
                 'name_en' => $request->name_en ,
                 'notes' => $request->notes ,
             ]);
+
             toastr()->success(trans('messages.success')) ;
             return redirect()->back();
         }
@@ -37,7 +37,6 @@ class GradesController extends Controller
         }
 
     }
-
 
 
     public function update(StoreGrades $request, $id)
@@ -62,14 +61,26 @@ class GradesController extends Controller
     }
 
 
-    public function destroy(Request $request)
+    public function destroy(Request $request , $id)
     {
 
         try {
-            $Grade = Grade::findOrFail($request->id)->delete();
-            toastr()->warning(trans('messages.Delete')) ;
-            return redirect()->back();
-        }
+
+            $Count_Classrooms = Classroom::where('grade_id'  , $request->id)->pluck('grade_id')->count();
+
+            if ($Count_Classrooms == 0) {
+                $Grade = Grade::findOrFail($request->id)->delete();
+                toastr()->warning(trans('messages.Delete')) ;
+                return redirect()->back();
+                 }
+
+            else {
+                toastr()->error(trans('Grades_trans.delete_Grade_Error')) ;
+                return redirect()->back();
+            }
+
+
+         }
 
         catch (\Exception $e){
             return redirect()->back()->withErrors(['error' => $e->getMessage()]);
